Sparate core impl and ROS node/nodelet wrapper
