+++ diagnostic_generic_diagnostics/include/diagnostic_generic_diagnostics/topic_monitor.h
@@ -1,3 +1,4 @@
+
 #include <ros/ros.h>
 #include <topic_tools/shape_shifter.h>
 #include "diagnostic_updater/update_functions.h"
@@ -158,18 +159,3 @@ public:
 };
 
 };  // namespace diagnostic_generic_diagnostics
-
-int main(int argc, char **argv)
-{
-  ros::init(argc, argv, "topic_monitor");
-
-  ros::NodeHandle nh("");
-  ros::NodeHandle pnh("~");
-
-  diagnostic_generic_diagnostics::TopicMonitor topic_monitor(nh, pnh);
-
-  ROS_INFO("spinning...");
-  ros::spin();
-  ROS_INFO("exit...");
-  return 0;
-}
